Move LibAssetDataTransfer from forwarder to exchange-libs package
