+++ contracts/broker/package.json
@@ -54,7 +54,6 @@
         "@0x/abi-gen": "^5.1.0",
         "@0x/contracts-erc721": "^3.0.4",
         "@0x/contracts-exchange": "^3.1.0",
-        "@0x/contracts-exchange-forwarder": "^4.0.4",
         "@0x/contracts-exchange-libs": "^4.1.0",
         "@0x/contracts-gen": "^2.0.4",
         "@0x/contracts-test-utils": "^5.1.1",
