+++ contracts/exchange-forwarder/CHANGELOG.json
@@ -1,5 +1,14 @@
 [
     {
+        "version": "4.1.0",
+        "changes": [
+            {
+                "note": "Moved LibAssetDataTransfer to exchange-libs",
+                "pr": 2455
+            }
+        ]
+    },
+    {
         "timestamp": 1580811564,
         "version": "4.0.5",
         "changes": [
