+++ contracts/exchange-forwarder/test/artifacts.ts
@@ -10,7 +10,6 @@ import * as IAssets from '../test/generated-artifacts/IAssets.json';
 import * as IExchangeV2 from '../test/generated-artifacts/IExchangeV2.json';
 import * as IForwarder from '../test/generated-artifacts/IForwarder.json';
 import * as IForwarderCore from '../test/generated-artifacts/IForwarderCore.json';
-import * as LibAssetDataTransfer from '../test/generated-artifacts/LibAssetDataTransfer.json';
 import * as LibConstants from '../test/generated-artifacts/LibConstants.json';
 import * as LibForwarderRichErrors from '../test/generated-artifacts/LibForwarderRichErrors.json';
 import * as MixinAssets from '../test/generated-artifacts/MixinAssets.json';
@@ -30,7 +29,6 @@ export const artifacts = {
     IExchangeV2: IExchangeV2 as ContractArtifact,
     IForwarder: IForwarder as ContractArtifact,
     IForwarderCore: IForwarderCore as ContractArtifact,
-    LibAssetDataTransfer: LibAssetDataTransfer as ContractArtifact,
     LibConstants: LibConstants as ContractArtifact,
     LibForwarderRichErrors: LibForwarderRichErrors as ContractArtifact,
     TestForwarder: TestForwarder as ContractArtifact,
