+++ packages/utils/src/index.ts
@@ -30,6 +30,7 @@ export {
 export import BrokerRevertErrors = require('./revert_errors/broker/revert_errors');
 export import CoordinatorRevertErrors = require('./revert_errors/coordinator/revert_errors');
 export import ExchangeForwarderRevertErrors = require('./revert_errors/exchange-forwarder/revert_errors');
+export import LibAssetDataTransferRevertErrors = require('./revert_errors/exchange-libs/lib_asset_data_transfer_revert_errors');
 export import LibMathRevertErrors = require('./revert_errors/exchange-libs/lib_math_revert_errors');
 export import ExchangeRevertErrors = require('./revert_errors/exchange/revert_errors');
 export import FixedMathRevertErrors = require('./revert_errors/staking/fixed_math_revert_errors');
